add is_valid funtion

+++ src/components/opinions_admin/OpinionAdminContainer.jsx
@@ -5,6 +5,7 @@ import './OpinionAdminContainer.scss'
 const OpinionAdminContainer = () => {
     const [opinions, setOpinions] = useState([])
     const [reload, setReload] = useState(false);
+    const [valid, setValid] = useState('à valider');
 
     useEffect(() => {
         const getOpinions = async () => {
@@ -13,7 +14,6 @@ const OpinionAdminContainer = () => {
                 .then((result) => result.data)
                 .then((data) => {
                     setOpinions(data);
-                    console.log(opinions);
                 })
                 .catch((err) => {
                     console.log(err);
@@ -27,33 +27,59 @@ const OpinionAdminContainer = () => {
         getOpinions();
     }, [reload]);
 
+    const handleUpdate = (idOpinion) => {
+        if(window.confirm('Voulez-vous valider cet avis ?')){
+            const url = `${process.env.REACT_APP_API_URL}opinion/${idOpinion}`;
+            axios.put(url, { is_valid: true }, { withCredentials: true })
+                .then((response) => {
+                    if (response.status === 200) {
+                        setReload(!reload);
+                    }
+                })
+                .catch((err) => {
+                    console.log(err.response);
+                    const HTTPError = err.response.status;
+                    if (HTTPError === 401) {
+                        alert('Vous avez été déconnecté.');
+                        window.location = '/';
+                    }
+                    if (HTTPError === 500 && err.response.data.error === 1451) {
+                        alert(err.response.data.message);
+                    }
+                });
+        }
+    };
+
     const handleRemove = (idOpinion) => {
-        const url = `${process.env.REACT_APP_API_URL}opinion/${idOpinion}`;
-        axios.delete(url, { withCredentials: true })
-            .then((response) => {
-                if (response.status === 200) {
-                    setReload(!reload);
-                }
-            })
-            .catch((err) => {
-                console.log(err.response);
-                const HTTPError = err.response.status;
-                if (HTTPError === 401) {
-                    alert('Vous avez été déconnecté.');
-                    window.location = '/';
-                }
-            });
+        if(window.confirm('Voulez-vous supprimer cet avis ?')){
+            const url = `${process.env.REACT_APP_API_URL}opinion/${idOpinion}`;
+            axios.delete(url, { withCredentials: true })
+                .then((response) => {
+                    if (response.status === 200) {
+                        setReload(!reload);
+                    }
+                })
+                .catch((err) => {
+                    console.log(err.response);
+                    const HTTPError = err.response.status;
+                    if (HTTPError === 401) {
+                        alert('Vous avez été déconnecté.');
+                        window.location = '/';
+                    }
+                });
+            }
     };
 
     return (
         <div className='OpinionAdminContainer'>
             {opinions.map((opinion) =>{
                 return(
-                    <div key={opinion.id_opinion}>
-                        <p>{opinion.author}</p>
-                        <p>{opinion.opinion}</p>
-                        <p>{opinion.is_valid}</p>
-                        <button onClick={() => handleRemove(opinion.id_opinion)}>Supprimer</button>
+                    <div key={opinion.id_opinion} className='opinionContainer'>
+                        <p className='opinionAuthor'>{opinion.author}</p>
+                        <p className='opinionDetail'>{opinion.opinion}</p>
+                        {opinion.is_valid ? <i className="fa-solid fa-check opinionIsValid"></i> : ''}
+                        {!opinion.is_valid ? <button onClick={() => handleUpdate(opinion.id_opinion)} className='opinionIsValidButton'>Valider l'avis</button> : ''}
+                        <button onClick={() => handleRemove(opinion.id_opinion)} className='opinionDeleteButton'>Supprimer l'avis</button>
                     </div>
                 )
             })}
